requirements.txt
loguru==0.7.2       # conda-forge
matplotlib==3.8.1   # pypi
numpy==1.22.4       # pypi
pandas==1.5.3       # pypi
rdkit==2023.9.2     # pypi
seaborn==0.13.2     # defaults
summit==0.8.9       # pypi
tqdm==4.65.0        # conda-forge
scikit-learn=1.5.1  # pypi
idaes-pse==1.5.1    # pypi
openpyxl=3.1.5      # pypi
umap-learn=0.5.6    # pypi
plotly=5.18.0       # pypi